hw8.taxi.action authentication

+++ vladimir_morochko_web/src/main/java/hw8/taxi/service/ClientServiceImpl.java
@@ -103,7 +103,7 @@ public class ClientServiceImpl implements ClientService {
     public List<Client> showClientsLastMonth() {
         List <Client> list = new ArrayList<>();
         Date currentDate = new Date();
-        Long monthMilliseconds = (long)1000 * 60 * 60 * 24 * 31;
+        Long monthMilliseconds = 1000L * 60 * 60 * 24 * 31;
         for (Client client : clients) {
             if (client.getLastOrderDate() != null
                     && (currentDate.getTime() - client.getLastOrderDate().getTime() < monthMilliseconds)) {
